requirements.txt
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
idna==3.10
mutagen==1.47.0
pycryptodomex==3.20.0
pydub==0.25.1
PyYAML==6.0.2
requests==2.32.3
termcolor==2.4.0
urllib3==2.2.3
websockets==13.1
yt-dlp==2024.9.27
